+++ src/Logitar.Portal.EntityFrameworkCore.PostgreSQL/Entities/ActorEntity.cs
@@ -21,6 +21,14 @@ internal record ActorEntity
   public static ActorEntity Deserialize(string json) => JsonSerializer.Deserialize<ActorEntity>(json, _options)
     ?? throw new InvalidOperationException($"The actor deserialization failed: '{json}'.");
 
+  public static ActorEntity From(Actor actor) => new()
+  {
+    Type = actor.Type,
+    IsDeleted = actor.IsDeleted,
+    DisplayName = actor.DisplayName,
+    EmailAddress = actor.EmailAddress,
+    Picture = actor.Picture
+  };
   public static ActorEntity From(UserEntity user, bool isDeleted = false) => new()
   {
     Type = ActorType.User,
@@ -31,4 +39,14 @@ internal record ActorEntity
   };
 
   public string Serialize() => JsonSerializer.Serialize(this, _options);
+
+  public Actor ToActor(Guid id) => new()
+  {
+    Id = id,
+    Type = Type,
+    IsDeleted = IsDeleted,
+    DisplayName = DisplayName,
+    EmailAddress = EmailAddress,
+    Picture = Picture
+  };
 }