+++ CHANGELOG.md
@@ -31,6 +31,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Reimplemented password recovery.
 - Implemented Basic authentication.
 - Reimplemented logging.
+- Reimplemented caching.
 
 ## [1.1.5] - 2022-10-27